feat(packaging): one `ggshield` command — Rust dispatcher for the hook, Python for the rest
The bundled `ggshield` becomes a small Rust dispatcher: it answers the exact
`secret scan ai-hook` argv itself, in about 20 ms instead of 320, and hands
every other command to the PyInstaller launcher beside it, now named
`ggshield-py`. There is still one `ggshield` on the PATH.

Dispatch is an exact match and nothing here parses ggshield's grammar, so
the only mistake it can make is handing an argv to Python. `exec` on Unix,
spawn-and-wait on Windows; argv, the standard streams and the exit code all
pass through untouched. A launcher that is missing, unrunnable, or turns out
to be this binary itself fails loudly rather than silently succeeding.

`build_hook_command()` writes the dispatcher's path on a bundle install;
pip, Homebrew and Docker are untouched.

The delegation suite covers argv, stdin/stdout/stderr, exit codes, symlink
invocation and a broken install. It spawns binaries the linker has only just
written, which on Linux earns `ETXTBSY` from a sibling thread's inherited
descriptor, so those spawns retry -- see the note there for how to make the
race fail on demand.
