+++ changelog.d/20260805_170000_achille.mascia_ggshield_rust_dispatcher.md
@@ -1,0 +1,6 @@
+### Changed
+
+- In the standalone packages (`.pkg`, `.deb`, `.rpm`, `.msi`, Chocolatey, archives), `ggshield`
+  is now a small native dispatcher: it runs `ggshield secret scan ai-hook` and hands every other
+  command to the bundled Python implementation. There is still a single `ggshield` command, and
+  every command behaves as before.
